Updated the Rake dependency to 11.1
In 599b295, the Rake dependency has been pinned to < 11, but that prevents installing
of gems depending on later versions. The offending `last_comment` method was restored in Rake 11.1,
so it is safe to pin that version (relaxed). Furthermore, a patch to Yard is coming up which
fixes the original problem there.

Thanks for reporting and all the pointers, @dominicsayers!

Closes #566

Related: lsegal/yard#964